backend/requirements.txt
Flask
Flask-Cors
PyPDF2
python-docx
scikit-learn
vercel-python